samd: Use MCU specific mpconfig_samdXX files.
These replace MCU-specific settings in mpconfigport.h and in
the mpconfigboard.h files.
